+++ composeml/label_maker.py
@@ -4,6 +4,7 @@ import pandas as pd
 from tqdm import tqdm
 
 from composeml.label_times import LabelTimes
+from composeml.offsets import to_offset
 from composeml.utils import can_be_type
 
 
@@ -55,48 +56,8 @@ def cutoff_data(df, threshold):
     return df, cutoff_time
 
 
-def is_offset(value):
-    """Checks whether a value is an offset.
-
-    Args:
-        value (any) : Value to check.
-
-    Returns:
-        Bool : Whether value is an offset.
-    """
-    return issubclass(type(value), pd.tseries.offsets.BaseOffset)
-
-
-def to_offset(value):
-    """Converts a value to an offset and validates the offset.
-
-    Args:
-        value (int or str or offset) : Value of offset.
-
-    Returns:
-        offset : Valid offset.
-    """
-    if isinstance(value, int):
-        assert value > 0, 'offset must be greater than zero'
-        offset = value
-
-    elif isinstance(value, str):
-        error = 'offset must be a valid string'
-        assert can_be_type(type=pd.tseries.frequencies.to_offset, string=value), error
-        offset = pd.tseries.frequencies.to_offset(value)
-        assert offset.n > 0, 'offset must be greater than zero'
-
-    else:
-        assert is_offset(value), 'invalid offset'
-        assert value.n > 0, 'offset must be greater than zero'
-        offset = value
-
-    return offset
-
-
 class Context:
     """Metadata for data slice."""
-
     def __init__(self, gap=None, window=None, slice_number=None, target_entity=None, target_instance=None):
         """Metadata for data slice.
 
@@ -137,7 +98,6 @@ class DataSlice(pd.DataFrame):
 
 class LabelMaker:
     """Automatically makes labels for prediction problems."""
-
     def __init__(self, target_entity, time_index, labeling_function, window_size=None, label_type=None):
         """Creates an instance of label maker.
 
@@ -323,7 +283,8 @@ class LabelMaker:
             minimum_data=minimum_data,
             gap=gap,
             drop_empty=drop_empty,
-            verbose=False)
+            verbose=False,
+        )
 
         name = self.labeling_function.__name__
         labels, instance = [], 0
